configure with text config from any source, not just from file
